Added class NewTextbox, dynamic expanding textbox
Class is unfinished, but the end goal is to have a dynamically
expanding textbox that will add a new row whenever I press enter,
but will also remove the bottom row if I backspace to where the
cursor is not on the bottom row and the bottom row is empty.

To be used as the primary input method in the lists.